- fix
git-svn-id: svn://dev-svn.seasr.org/Components/Foundry/trunk@2499 dc7ab104-3778-4984-a5d9-166996b8e16d

+++ src-vis/org/seasr/meandre/components/vis/html/FrameMaker.java
@@ -195,7 +195,7 @@ public class FrameMaker extends AbstractStreamingExecutableComponent {
         VelocityContext context = velocity.getNewContext();
 
         context.put("htmlDocs", _htmlDocs);
-        context.put("rows", (int)Math.ceil(_htmlDocs.size() / _columns));
+        context.put("rows", (int)Math.ceil(_htmlDocs.size() / (double)_columns));
 		context.put("columns", Math.min(_columns, _htmlDocs.size()));
 
         String html = velocity.generateOutput(context, _template);
